# perf(scripts-task): turn off path aliases within generate-api task, make isUsingPathAliasesForDx lazy => improve d.ts generation speed by ~60%