--- a/change/@fluentui-react-field-b0fef346-1748-4494-95c1-40003abf4b2f.json
+++ b/change/@fluentui-react-field-b0fef346-1748-4494-95c1-40003abf4b2f.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "chore: re-genearate api.md as result of perf improvements to generate-api task",
+  "packageName": "@fluentui/react-field",
+  "email": "martinhochel@microsoft.com",
+  "dependentChangeType": "none"
+}