<!--
Thank you for submitting a pull request!

Please verify that:
* [ ] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [ ] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->

### Pre-requirements:

- [x] https://github.com/microsoft/fluentui/pull/30989

## Previous Behavior

<!-- This is the behavior we have today -->

## New Behavior

- `isUsingPathAliasesForDx` us not used anymore for v9 packages - refactored to make it lazy for few v0/v8 invocations
- TS path aliases are not used anymore for `d.ts` generation in v9 packages - which yields significant perf boost ( **up to 60%  faster** - based on package )

| command | Before | After |
|--------|--------|--------|
| yarn workspace @fluentui/react-drawer generate-api | 16s | 7s /  **𝚫 56% faster** |


## Related Issue(s)

<!-- Please link the issue being fixed so it gets closed when this is merged. -->

- Partially implements https://github.com/microsoft/fluentui/issues/30516