@fabricteam 2024-04-08
## 📊 Bundle size report

✅ No changes found
<div id="bundleSizeComment9423"/>